+++ saleor/plugins/webhook/tests/test_utils.py
@@ -4,13 +4,12 @@ import pytest
 
 from ....payment.interface import PaymentGateway
 from ....webhook.event_types import WebhookEventSyncType
-from ..utils import (
-    generate_cache_key_for_webhook,
+from ..list_stored_payment_methods import (
     get_credit_card_info,
     get_list_stored_payment_methods_from_response,
     get_payment_method_from_response,
-    to_payment_app_id,
 )
+from ..utils import generate_cache_key_for_webhook, to_payment_app_id
 
 
 @pytest.fixture()
